将滥用的unsigned int等内置类型改为定长的uint32_t等类型
C/C++中的int不一定是4字节的。C99规定了uint32_t等定长的数据类型，这样就能方便地
处理类似Multiboot引导信息的结构了。

一些无需担心数据在内存中的长度的变量的类型没有改变，这包括pos和rect等结构中的
x、y、width、height等。

+++ core/terminal.cpp
@@ -56,7 +56,7 @@ namespace Terminal {
 	void set_pos(struct pos pos) {
 		cursor_x = pos.x;
 		cursor_y = pos.y;
-		unsigned short pos_info = pos.x + pos.y * width;
+		uint8_t pos_info = pos.x + pos.y * width;
 		asm_out8(0x03d4, 14);
 		asm_out8(0x03d5, pos_info >> 8);
 		asm_out8(0x03d4, 15);
